cleanup comments
